+++ docs/source/whats_new.rst
@@ -17,10 +17,12 @@ Develop branch
 
 Enhancements
 ~~~~~~~~~~~~
+- Add possibility to use OptunaGridSearch (:gh:`630` by `Igor Carrara`_)
 - Add scripts to upload results on PapersWithCode (:gh:`561` by `Pierre Guetschel`_)
 - Centralize dataset summary tables in CSV files (:gh:`635` by `Pierre Guetschel`_)
 - Add new dataset :class:`moabb.datasets.Liu2024` dataset (:gh:`619` by `Taha Habib`_)
 - Increasing the version in the pre-commit config (:gh:`631` by pre-commit bot)
+
 
 
 Bugs
@@ -29,7 +31,9 @@ Bugs
 
 API changes
 ~~~~~~~~~~~
-- None
+- Include optuna as soft-dependency in the benchmark function and in the base of evaluation (:gh:`630` by `Igor Carrara`_)
+
+
 
 Version - 1.1.0  (Stable - PyPi)
 ---------------------------------